Adaptive write delay
This commit adds the ability to use a rudimentary adaptive write delay
when writing data with inch.

If the `-target-latency` flag is set, then `inch` will attempt to
increase writers' delays when the server responses take longer than
`target-latency`, and reduce the writers' delays when they're quicker
than `target-latency`.

In general, inch adapts the current delay proportionally to the number
of concurrent writers and the size of the difference between the server
response time and the desired maximum time.