Add support for writing across virtual hosts
This commits adds a new flag `-vhosts`, which when set will uniformly
spread batches across different "virtual hosts". All writes continue
going to the host specified with the `-host` flag, but an
`X-Influxdb-Host` header is set when virtual hosts are specified.

This feature is for InfluxData testing of internal systems.