feat: allow one write per field for each point
Add a flag -multiwrite that spreads the writes
for each point out to one write per field,
while holding the timestamp and tags the same.
This permits testing the cost of handling a
write as distinct from the cost of saving a
point

Also fixes points in a series overwriting each
other because of identical timestamps.