feat: update inch options and url to support influxdb 3 core
Adds an option to use the v3 native write endpoint so advanced control
of write behavior can be used such as the "no-sync" option.

Additionally, modifies the channel for buffered write batches to reflect
concurrency setting, fixes an issue with retrying a write that was
needlessly recompressing or copying the write batch, reduced allocations
and copies by reusing buffers and letting bytes Buffer's take ownership
of existing buffers.

Updates golang to version 1.24

* closes #46